Run tests against multiple versions of GraphQL.js
