+++ __tests__/graphql-pr-4192.test.mjs
@@ -1,0 +1,3 @@
+import { runTest } from "./runTest.mjs";
+
+runTest("graphql-pr-4192");
