+++ __tests__/graphql15.test.mjs
@@ -1,0 +1,3 @@
+import { runTest } from "./runTest.mjs";
+
+runTest("graphql15");
