+++ __tests__/graphql17.test.mjs
@@ -1,0 +1,3 @@
+import { runTest } from "./runTest.mjs";
+
+runTest("graphql17");
